pyproject.toml
[project]
name = "brewmaster"
version = "0.1.0"
description = "Agent for Bees-Bewery-Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jldracula>=0.2.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "python-dotenv>=1.2.0",
    "ipykernel>=7.1.0",
    "langchain-anthropic>=1.0.3",
    "langchain-model-profiles>=0.0.4",
    "langchain-tavily>=0.2.13",
    "tavily-python>=0.7.13",
    "tavily>=1.1.0",
    "langchain-community>=0.4.1",
    "pypdf>=6.2.0",
    "langchain-text-splitters>=1.0.0",
    "mcp>=1.21.1",
    "langchain-mcp-adapters>=0.1.13",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.9",
    "langgraph-api>=0.5.37",
    "ipywidgets>=8.1.8",
    "sounddevice>=0.5.3",
    "scipy>=1.16.3",
    "langchain-google-vertexai>=3.0.3",
    "langchain-google-genai>=3.1.0",
    "langsmith>=0.4.43",
    "langchain-core>=1.1.3",
    "faker>=40.15.0",
    "chromadb>=1.5.8",
    "langchain-chroma>=1.1.0",
    "rank-bm25>=0.2.2",
]
